# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.updateMeRequest:
    properties:
      avatar_url:
        type: string
      name:
        type: string
    type: object
  handler.updateMetadataRequest:
    properties:
      metadata_private:
        additionalProperties: true
        type: object
      metadata_public:
        additionalProperties: true
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@saaskit.io
    name: SaaSKit Support
    url: https://github.com/medaminerjb/saas-kit/issues
  description: Enterprise-grade Identity and Access Management platform with OIDC,
    multi-tenant support, and API key management.
  license:
    name: MIT
    url: https://github.com/medaminerjb/saas-kit/blob/main/LICENSE
  termsOfService: https://github.com/medaminerjb/saas-kit/blob/main/TERMS.md
  title: SaaSKit API
  version: "1.0"
paths:
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update the authenticated user's profile
      parameters:
      - description: Profile updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Update failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update current user
      tags:
      - users
  /users/me/metadata:
    get:
      consumes:
      - application/json
      description: Get public metadata for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User metadata
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user metadata
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update metadata for the authenticated user
      parameters:
      - description: Metadata updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateMetadataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated metadata
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Update failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user metadata
      tags:
      - users
  /users/me/sessions:
    get:
      consumes:
      - application/json
      description: Get all active sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of sessions
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List user sessions
      tags:
      - users
  /users/me/sessions/{sessionID}:
    delete:
      consumes:
      - application/json
      description: Revoke a specific session for the authenticated user
      parameters:
      - description: Session ID
        in: path
        name: sessionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session revoked
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid session ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Revoke failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Revoke a session
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"